requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart
bcrypt
# email-validator
pyjwt
